Simplify codegen more
